pyproject.toml
[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119

[tool.poetry]
name = "proactive-voice-agent"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["uvicorn"], version = "^0.110.2"}
mistralai = "^0.1.8"
langchain-community = "^0.0.34"
uvicorn = "^0.29.0"
retell-sdk = "^3.11.0"
sentence-transformers = "^2.7.0"
faiss-cpu = "^1.8.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
